chore: project restructuring feat: create daily data update workflow
