+++ package.json
@@ -3,7 +3,9 @@
   "module": "src/index.ts",
   "type": "module",
   "scripts": {
-    "data": "bun run src/index.ts"
+    "fetch-plugin-data": "bun run src/plugin/index.ts",
+    "fetch-theme-data": "bun run src/theme/index.ts",
+    "fetch-release-data": "bun run src/release/index.ts"
   },
   "devDependencies": {
     "bun-types": "latest"
